refactor: split test cases into multiple files
